+++ src/platform/APM32/dshot_bitbang.c
@@ -665,7 +665,7 @@ static void bbShutdown(void)
     return;
 }
 
-static bool bbIsMotorEnabled(uint8_t index)
+static bool bbIsMotorEnabled(unsigned index)
 {
     return bbMotors[index].enabled;
 }
