+++ src/platform/AT32/dshot_bitbang.c
@@ -657,7 +657,7 @@ static void bbShutdown(void)
     return;
 }
 
-static bool bbIsMotorEnabled(uint8_t index)
+static bool bbIsMotorEnabled(unsigned index)
 {
     return bbMotors[index].enabled;
 }
